requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
cycler==0.10.0
discord.py==1.7.3
dnspython==1.16.0
idna==3.2
kiwisolver==1.3.1
loguru==0.5.3
matplotlib==3.4.2
multidict==5.1.0
numpy==1.21.5
Pillow==8.3.1
pydantic==1.8.2
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.0
typish==1.9.2
tzlocal==2.1
urllib3==1.26.6
yarl==1.6.3
